Privileges

+++ controller/UsersController.js
@@ -1,24 +1,16 @@
 const Users = require('../models/Users');
 const CPF = require('cpf');
-
-function __render_with_message(req, res, page, variables_of_page, message) {
-    req.flash('message', message);
-    res.render(page, variables_of_page);
-}
-
-function __redirect_with_message(req, res, url, message) {
-    req.flash('message', message);
-    res.redirect(url);
-}
+const RolesController = require('./RolesController');
+const Utils = require('./Utils');
 
 class UserController {
     static login(req, res) {
         try {
             if (req.session.userid) {
-                __redirect_with_message(req, res, '/administracao/painel', 'Você já esta autenticado');
+                Utils.Message.redirect(req, res, '/administracao/painel', 'Você já esta autenticado');
                 return
             }
-            res.render('login', { 'title': 'Autenticação' });
+            res.render(Utils.Titles.login.name_page, { 'title': Utils.Titles.login.title });
         } catch (error) {
             console.log(error)
         }
@@ -31,17 +23,17 @@ class UserController {
                 where: { email: email }
             });
             if (!user) {
-                __render_with_message(req, res, 'login', { 'title': 'Autenticação' }, 'Usuário não cadastrado');
+                Utils.Message.render(req, res, Utils.Titles.login.name_page, { 'title': Utils.Titles.login.title }, 'Usuário não cadastrado');
                 return
             }
-            if (!password === user.password) {
-                __render_with_message(req, res, 'login', { 'title': 'Autenticação' }, 'Senha invalida');
+            if (password !== user.password) {
+                Utils.Message.render(req, res, Utils.Titles.login.name_page, { 'title': Utils.Titles.login.title }, 'Senha invalida');
                 return
             }
             req.session.userid = user.id;
-            req.session.userrole = user.role;
+            req.session.userrole = user.RoleId;
             req.session.save(() => {
-                __redirect_with_message(req, res, '/administracao/painel', 'Sejá bem vindo(a) ' + user.full_name + '.');
+                Utils.Message.redirect(req, res, '/administracao/painel', 'Sejá bem vindo(a) ' + user.full_name + '.');
             })
         } catch (error) {
             console.log(error);
@@ -75,9 +67,9 @@ class UserController {
         }
     }
 
-    static addUser(req, res) {
-        if (req.session.userid && req.session.userrole === 'adiministrador') {
-            res.render('addUser', { 'title': 'Adicionar novo funcionário' });
+    static async addUser(req, res) {
+        if (RolesController.isAdmin(req.session.userrole)) {
+            res.render(Utils.Titles.addUser.name_page, { 'title': Utils.Titles.addUser.title, 'allRoles': await RolesController.allRoles() });
             return
         }
         res.redirect('/administracao');
@@ -85,31 +77,32 @@ class UserController {
 
     static async addUserPost(req, res) {
         try {
-            if (req.session.userid && req.session.userrole === 'adiministrador') {
-                let { full_name, cpf, email, password, role } = req.body;
-                if (!CPF.isValid(cpf)) {
-                    __render_with_message(req, res, 'addUser', { 'title': 'Adicionar novo funcionário' }, 'CPF invalido.');
-                    return;
-                }
-                cpf = CPF.format(cpf);
-                const user_test_cpf_exists = await Users.findOne({ where: { cpf: cpf } });;
-                const user_test_email_exists = await Users.findOne({ where: { email: email } });;
-                if (user_test_cpf_exists || user_test_email_exists) {
-                    __render_with_message(req, res, 'addUser', { 'title': 'Adicionar novo funcionário' }, 'Usuário já cadastrado.');
-                    return;
-                }
-                const new_user = await Users.create({
-                    full_name: full_name,
-                    cpf: cpf,
-                    email: email,
-                    password: password,
-                    role: role
-                });
-                new_user.save();
-                __redirect_with_message(req, res, '/administracao/adicionar_funcionario', 'Funcionario adicionado com sucesso.')
+            if (!RolesController.isAdmin(req.session.userrole)) {
+                res.redirect('/administracao');
                 return
             }
-            res.redirect('/administracao');
+            let { full_name, cpf, email, password, RoleId } = req.body;
+            if (!CPF.isValid(cpf)) {
+                Utils.Message.render(req, res, Utils.Titles.addUser.name_page, { 'title': Utils.Titles.addUser.title }, 'CPF invalido.');
+                return;
+            }
+            cpf = CPF.format(cpf);
+            const user_test_cpf_exists = await Users.findOne({ where: { cpf: cpf } });;
+            const user_test_email_exists = await Users.findOne({ where: { email: email } });;
+            if (user_test_cpf_exists || user_test_email_exists) {
+                Utils.Message.render(req, res, Utils.Titles.addUser.name_page, { 'title': Utils.Titles.addUser.title }, 'Usuário já cadastrado.');
+                return;
+            }
+            const new_user = await Users.create({
+                full_name: full_name,
+                cpf: cpf,
+                email: email,
+                password: password,
+                RoleId: RoleId
+            });
+            new_user.save();
+            Utils.Message.redirect(req, res, '/administracao/adicionar_funcionario', 'Funcionario adicionado com sucesso.')
+            return
         } catch (error) {
             console.log(error)
         }
@@ -120,7 +113,7 @@ class UserController {
             let { full_name, cpf, email, password } = req.body;
             if (!CPF.isValid(cpf)) {
                 const user = await Users.findOne({ where: { id: req.session.userid } });
-                __render_with_message(req, res, 'usuario', {
+                Utils.Message.render(req, res, 'usuario', {
                     title: user.full_name,
                     full_name: user.full_name,
                     cpf: user.cpf,
@@ -138,7 +131,7 @@ class UserController {
                 },
                 { where: { id: req.session.userid } }
             );
-            __redirect_with_message(req, res, '/administracao/usuario', 'Alteração efetuada com sucesso.')
+            Utils.Message.redirect(req, res, '/administracao/usuario', 'Alteração efetuada com sucesso.')
         } catch (error) {
             console.log(error);
         }
@@ -146,7 +139,7 @@ class UserController {
 
     static async removeUser(req, res) {
         try {
-            if (req.session.userid && req.session.userrole === 'adiministrador') {
+            if (RolesController.isAdmin(req.session.userrole)) {
                 const allUsers = await Users.findAll();
                 var allUsersSimplified = [];
                 allUsers.forEach((id) => {
@@ -159,7 +152,7 @@ class UserController {
                         );
                     }
                 });
-                res.render('removeUser', { 'title': 'Remover funcionário', 'allUsers': allUsersSimplified });
+                res.render(Utils.Titles.removeUser.name_page, { 'title': Utils.Titles.removeUser.title, 'allUsers': allUsersSimplified });
                 return
             }
             res.redirect('/administracao');
@@ -170,15 +163,15 @@ class UserController {
 
     static async removeUserPost(req, res) {
         try {
-            if (req.body.select_employee !== req.session.userid) {
+            if (req.body.select_employee !== req.session.userid && RolesController.isAdmin(req.session.userrole)) {
                 await Users.destroy({
                     where: {
                         id: req.body.select_employee
                     },
                 });
-                __redirect_with_message(req, res, '/administracao/remover_funcionario', 'Funcionario removido com sucesso.');
+                Utils.Message.redirect(req, res, '/administracao/remover_funcionario', 'Funcionario removido com sucesso.');
             }
-            __redirect_with_message(req, res, '/administracao/remover_funcionario', 'Você tentou remover sua propria conta.');
+            Utils.Message.redirect(req, res, '/administracao/remover_funcionario', 'Você tentou remover sua propria conta.');
         } catch (error) {
             console.log(error);
         }
@@ -186,20 +179,20 @@ class UserController {
 
     //a fazer
     static async functionalities(req, res) {
-        res.render('functionalities', { 'title': 'Funcionalidades' });
+        res.render(Utils.Titles.functionalities.name_page, { 'title': Utils.Titles.functionalities.title });
     }
 
     //a fazer
     static async panel(req, res) {
-        res.render('panel', { 'title': 'Painel' });
+        res.render(Utils.Titles.panel.name_page, { 'title': Utils.Titles.panel.title });
     }
 
     //a fazer
     static async panelPost(req, res) {
-        res.render('panel', { 'title': 'Painel' });
+        res.render(Utils.Titles.panel.name_page, { 'title': Utils.Titles.panel.title });
     }
 
-    static async __addAdmin() {
+    static async setAdmin() {
         try {
             await Users.findOrCreate({
                 where: { id: 1 },
@@ -208,7 +201,8 @@ class UserController {
                     cpf: '000.000.000-00',
                     email: 'admin@email.com',
                     password: 'admin',
-                    role: 'adiministrador'
+                    role: 'adiministrador',
+                    RoleId: Utils.Roles.admin.id
                 }
             });
         } catch (error) {
@@ -216,7 +210,5 @@ class UserController {
         }
     }
 }
-
-UserController.__addAdmin();
 
 module.exports = UserController;
